Fix int division bug when computing chunks for each dimension
The calls to Math.ceil were effectively no-ops, as int/int produces an
int, which was then passed to the ceil method.

+++ cdm/zarr/src/main/java/ucar/nc2/iosp/zarr/ZarrHeader.java
@@ -397,7 +397,7 @@ public class ZarrHeader {
       int[] shape = zarray.getShape();
       int[] chunkSize = zarray.getChunks();
       for (int i = 0; i < nDims; i++) {
-        nChunks[i] = (int) Math.ceil(shape[i] / chunkSize[i]);
+        nChunks[i] = (int) Math.ceil((double) shape[i] / chunkSize[i]);
       }
       return ZarrUtils.subscriptsToIndex(subs, nChunks);
     } else {
